Update Resque dependency to 1.22.0.
Loner feature needs it to work when running Resque inline. In that Resque version, around_perform_hooks are enabled for inline mode. Before, an inline job would run and leave its lock behind, preventing new jobs with the same identifier from being enqueued.

Add a test that fails with an older version of Resque and passes with 1.22.0.